Boss Implemented
The boss type enemy has been implemented. It spawns with the triple shot weapon with a unique green colour to differentiate it from the player's own triple shot, as well as to differentiate it from the spam of enemy red bullets.

The enemy is also a purple in contrast to the other red enemy's, and has a scale factor of 1.75x so it is immediately identifiable as a more dangerous enemy to the player.

The vector math for the bullet prefab has also been adjusted so that the triple shot will fire downwards for the enemy, and still remain shooting upwards for the player